+++ src/module.rs
@@ -1,0 +1,1 @@
+//! Functions and structures that can be used to implement a PAM module