+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add Cargo features for `auth` (default) and `module`
 - Add internal `pam-macros` crate (experimental for now)
+- Reexport `pam_sys` at crate root
 
 ### Changed
 - Change `Authenticator::get_handler` to `Authenticator::handler_mut` and add `Authenticator::handler` for immutable access to the handler